requirements.txt
addict==2.4.0
certifi==2022.9.24
charset-normalizer==2.1.1
clip==1.0
contourpy==1.0.5
cycler==0.11.0
efficientnet-pytorch==0.7.1
fonttools==4.37.3
ftfy==6.1.1
gensim==4.2.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.0
mmcls==0.23.2
mmcv==1.6.2
mmsegmentation==0.28.0
munch==2.5.0
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pretrainedmodels==0.7.4
prettytable==3.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scipy==1.9.1
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@1fa49d0953708c0a61ab11c5cc4052d20909a9bc
six==1.16.0
smart-open==6.2.0
timm==0.4.12
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
yapf==0.32.0
